feat(user): Обновлены некоторые реализации в USER
