+++ taskforce/apps/task/src/app/app.module.ts
@@ -1,10 +1,9 @@
 import { Module } from '@nestjs/common';
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
+import { TaskModule } from './task/task.module';
 
 @Module({
-  imports: [],
-  controllers: [AppController],
-  providers: [AppService],
+  imports: [TaskModule],
+  controllers: [],
+  providers: [],
 })
 export class AppModule {}
