+++ taskforce/apps/task/src/app/task/dto/create-task.dto.ts
@@ -1,0 +1,17 @@
+export class CreateTaskDto {
+  title: string;
+
+  description: string;
+
+  category: string;
+
+  price?: number;
+
+  lifeTime?: Date;
+
+  image?: string;
+
+  address?: string;
+
+  tags?: string[];
+}
