+++ taskforce/apps/task/src/app/task/task.controller.ts
@@ -1,4 +1,4 @@
 import { Controller } from '@nestjs/common';
 
 @Controller('task')
-export class AppController {}
+export class TaskController {}
