+++ taskforce/apps/task/src/app/task/task.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { TaskController } from './task.controller';
+import { TaskService } from './task.service';
+
+@Module({
+  controllers: [TaskController],
+  providers: [TaskService],
+})
+export class TaskModule {}
