+++ taskforce/apps/task/src/app/task/task.service.ts
@@ -1,4 +1,4 @@
 import { Injectable } from '@nestjs/common';
 
 @Injectable()
-export class AppService {}
+export class TaskService {}
