+++ taskforce/apps/user/src/app/auth/auth.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
 import { UserMemoryModule } from '../user-memory/user-memory.module';
+import { UserModule } from '../user/user.module';
 import { AuthController } from './auth.controller';
-import { AuthService } from './auth.service';
 
 @Module({
-  imports: [UserMemoryModule],
+  imports: [UserMemoryModule, UserModule],
   controllers: [AuthController],
-  providers: [AuthService],
+  providers: [],
 })
 export class AuthModule {}
