+++ taskforce/apps/user/src/app/user/user.module.ts
@@ -7,5 +7,6 @@ import { UserService } from './user.service';
   imports: [UserMemoryModule],
   controllers: [UserController],
   providers: [UserService],
+  exports: [UserService]
 })
 export class UserModule {}
